pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wqv-viewer"
version = "0.1.0"
description = "PyQt6 desktop viewer for Casio WQV wrist camera images"
authors = [{name = "Community"}]
dependencies = [
    "Pillow>=9.0",
    "PyQt6>=6.5",
    "torch>=2.1",
    "realesrgan>=0.3.0",
    "numpy>=1.24",
    "basicsr>=1.4.2",
    "facexlib>=0.2.5",
    "gfpgan>=1.3.5",
    "opencv-python>=4.5",
    "torchvision>=0.16",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
]

[tool.setuptools]
packages = ["wqv_viewer"]
